refractor: moved all enums to a singular folder
