+++ src/main/java/de/cytooxien/realms/api/enums/WorldPreset.java
@@ -1,4 +1,4 @@
-package de.cytooxien.realms.api.world;
+package de.cytooxien.realms.api.enums;
 
 /**
  * Date: 08.10.2021
